code/home/pi/Smart-Security-Camera/requirements.txt
imutils
flask
picamera[array]
RPi.GPIO